% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devhelp-package.R
\docType{package}
\name{devhelp-package}
\alias{devhelp}
\alias{devhelp-package}
\title{devhelp: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Jennifer Bryan \email{jenny@posit.co} (\href{https://orcid.org/0000-0002-6983-2759}{ORCID})

}
\keyword{internal}
