% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/garply.R
\name{weeble}
\alias{weeble}
\alias{garply}
\title{A boring function}
\usage{
garply(x)
}
\arguments{
\item{x}{A character vector}
}
\value{
A character vector
}
\description{
The point of this topic is to give a topic name different from the function
name. Two other nifty functions in this pointless package are \code{\link[=foofy]{foofy()}}
and \code{\link[=blarg]{blarg()}}.
}
\examples{
garply()
}
