% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devhelp-package.R
\docType{package}
\name{devhelp-package}
\alias{devhelp}
\alias{devhelp-package}
\title{devhelp: What the Package Does (One Line, Title Case)}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jennybc/devhelp}
  \item Report bugs at \url{https://github.com/jennybc/devhelp/issues}
}

}
\author{
\strong{Maintainer}: Jennifer Bryan \email{jenny@posit.co} (\href{https://orcid.org/0000-0002-6983-2759}{ORCID})

}
\keyword{internal}
